fix(types): Typings Optimization
