+++ .eslintrc.yml
@@ -10,6 +10,7 @@ globals:
   BigInt: readonly
 parserOptions:
   ecmaVersion: 2020
+  sourceType: module
 rules:
   indent:
     - error
