--- types/libraries/Libraries.d.ts
@@ -1,2 +1,0 @@
-export { DiscordJS } from './wrappers/DiscordJS';
-export { Eris } from './wrappers/Eris';
