+++ types/libraries/index.d.ts
@@ -1,0 +1,2 @@
+export { DiscordJS } from "./DiscordJS";
+export { Eris } from "./Eris";
