--- types/node/ShoukakuRouter.d.ts
@@ -1,3 +1,0 @@
-import { ShoukakuRest } from './ShoukakuRest';
-
-export function ShoukakuRouter(rest: ShoukakuRest): typeof Proxy | Object | string | void;
