--- types/struct/ShoukakuTrackList.d.ts
@@ -1,9 +1,0 @@
-import { ShoukakuTrack } from './ShoukakuTrack';
-
-export class ShoukakuTrackList {
-    constructor(raw: object);
-    public type: 'PLAYLIST' | 'TRACK' | 'SEARCH' | 'NO_MATCHES' | 'LOAD_FAILED';
-    public selectedTrack: number;
-    public playlistName?: string;
-    public tracks: ShoukakuTrack[];
-}
